ordering orders

+++ taxiapp/models.py
@@ -36,7 +36,7 @@ class Meal(models.Model):
     name = models.CharField(max_length=500)
     short_description = models.CharField(max_length=500)
     image = models.ImageField(upload_to='meal_images/', blank=False)
-    price = models.IntegerField(default=0)
+    price = models.FloatField(default=0)
 
     def __str__(self):
         return self.name
@@ -58,7 +58,7 @@ class Order(models.Model):
     restaurant = models.ForeignKey(Restaurant, on_delete=models.CASCADE,)
     driver = models.ForeignKey(Driver, on_delete=models.CASCADE, blank = True, null = True)
     address = models.CharField(max_length=500)
-    total = models.IntegerField()
+    total = models.FloatField()
     status = models.IntegerField(choices = STATUS_CHOICES)
     created_at = models.DateTimeField(default = timezone.now)
     picked_at = models.DateTimeField(blank = True, null = True)
@@ -71,7 +71,7 @@ class OrderDetails(models.Model):
     order = models.ForeignKey(Order, on_delete=models.CASCADE, related_name='order_details')
     meal = models.ForeignKey(Meal, on_delete=models.CASCADE,)
     quantity = models.IntegerField()
-    sub_total = models.IntegerField()
+    sub_total = models.FloatField()
 
     def __str__(self):
         return str(self.id)
